refactor(db): extract reminder delivery store
